component file structure refactor
