+++ src/components/Core/Main.tsx
@@ -5,8 +5,8 @@ import { CookieSetOptions } from "universal-cookie";
 import { ApolloClient, NormalizedCacheObject } from "@apollo/client";
 import { Box, Container } from "@mui/material";
 
-import { Drawer, Navbar } from ".";
-import { JWT } from "../models";
+import { Drawer, Navbar } from "..";
+import { JWT } from "../../models";
 
 type props = {
   theme: string;
