+++ src/components/index.tsx
@@ -1,16 +1,16 @@
-import Admin from "./Admin";
-import ConfigField from "./ConfigField";
-import CreateCheckModal from "./CreateCheckModal";
-import CreateUserModal from "./CreateUserModal";
-import DeleteCheckModal from "./DeleteCheckModal";
-import DeleteUserModal from "./DeleteUserModal";
-import DrawerComponent from "./Drawer";
-import EditCheck from "./EditCheck";
-import EditUser from "./EditUser";
-import Error from "./Error";
-import Main from "./Main";
-import Navbar from "./Navbar";
-import PasswordInput from "./PasswordInput";
+import Admin from "./Admin/Admin";
+import ConfigField from "./Admin/Checks/ConfigField";
+import CreateCheckModal from "./Admin/Checks/CreateCheckModal";
+import CreateUserModal from "./Admin/Users/CreateUserModal";
+import DeleteCheckModal from "./Admin/Checks/DeleteCheckModal";
+import DeleteUserModal from "./Admin/Users/DeleteUserModal";
+import DrawerComponent from "./Core/Drawer";
+import EditCheck from "./Admin/Checks/EditCheck";
+import EditUser from "./Admin/Users/EditUser";
+import Error from "./Core/Error";
+import Main from "./Core/Main";
+import Navbar from "./Core/Navbar";
+import PasswordInput from "./Core/PasswordInput";
 
 export {
   Admin,
